backend/requirements.txt
asgiref==3.8.1
attrs==25.3.0
boto3==1.42.2
botocore==1.42.2
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.3
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==4.1
django-filter==25.1
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
gunicorn==23.0.0
idna==3.11
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
packaging==25.0
pillow==11.3.0
psycopg2_binary==2.9.11
PyJWT==2.9.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.1
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.11.0
pywebpush==2.0.0
django-apscheduler==0.7.0
apscheduler==3.10.4
cryptography>=3.4.8